+++ site/admin/2fa.php
@@ -1,18 +1,22 @@
 <?php
   session_start();
+
   include_once "configuration.php";
   include_once $USOC["SITE_PATH"]."/includes/class.inc.php";
-  include_once 'src/FixedBitNotation.php';
-  include_once 'src/GoogleAuthenticatorInterface.php';
-  include_once 'src/GoogleAuthenticator.php';
+  include_once "src/FixedBitNotation.php";
+  include_once "src/GoogleAuthenticatorInterface.php";
+  include_once "src/GoogleAuthenticator.php";
   include_once 'src/GoogleQrUrl.php';
+
+  newClass();
+
   $g = new \Sonata\GoogleAuthenticator\GoogleAuthenticator();
   if(isset($_POST["secret"])&&isset($_POST["register"])){
     $secret = $_POST["secret"];
     $db_link = mysqli_connect(MYSQL_HOST,MYSQL_USER,MYSQL_PASSWORD,MYSQL_DATABASE);
     $sql = "UPDATE User SET google_2fa='".$secret."' WHERE Username='".$_SESSION["User_Name"]."'";
     $db_erg = mysqli_query( $db_link, $sql );
-    echo "Google Authenticator verknüpft";
+    echo str_replace("%a",$U->getLang("login.2fa.google_authenticator"),$U->getLang("login.2fa.succeed"));
   }elseif(isset($_POST["code"])&&isset($_POST["login"])){
     $code = $_POST["code"];
     $db_link = mysqli_connect(MYSQL_HOST,MYSQL_USER,MYSQL_PASSWORD,MYSQL_DATABASE);
@@ -34,4 +38,4 @@
     $db_erg = mysqli_query( $db_link, $sql );
     header("Location: /index.php");
   }
- ?>
+?>